add version for swap encoding
When the new version of meson contracts are deployed, the same signatures can be executed on both old and new contracts. Added a version to swap encoding to prevent this issue in the future. Encoding with versions cannot be executed on the current deployed meson contracts because amount will overflow.